SBTS history assignment: Update formatting for Church History II
For CH2, all assignments are to be handed in together, so change heading such
that the lecture is put in a box. And because some assignments cover multiple
lectures while others address a single lecture, do not use automatic numbering,
but the lecture number will be included as part of the assignment title.