Change name of the script which updates my user packages and classes
Rename installation/update script so as to be more explicit about what it does.
The old name "updateinstall" was fine when being called explicitly from the
directory in which I maintain my classes and packages. But I want to be able to
call the script from other directories often (if compiling a document elsewhere,
I don't want to maintain two active directories to switch back and forth
between). Being explicitly about TeX in its name also makes it easier to find
though command line completion, for "updateinstall", I have to remember what
it's called. For "texupdate" if I start typing tex then it's easy to figure out
which of the possible commands I want.