Update paratext.sty to handle RTL input.
When using paratext for Hebrew, the mark settings were breaking the formatting
of the header references, putting elements the wrong way round. This was
identified as being due to the mark setting commands setting with RTL text.
Converting the input of \vn and \cn commands to LTR fixes the problem.