Fix size of header text in hbms class
Since changing size of Hebrew text in hbms class, the headers displaying verse
and page information have also been scaled up. Setting them to english text,
using the \textenglish macro from the polyglossia package and fontspec.

This will cause problems if non-English text is required to be displayed in the
headers. But for displaying verse and page number information, there are no
issues.