Update paratext and Greek/Hebrew classes to use new mark module
In 2022 release of LaTeX, a new mark module has been added to LaTeX, the
`ltmarks` module. This brings two major improvements for my usage over the
previous TeX marks functionality with various added LaTeX classes. Firstly, it
allows naming marks (and having an arbitrary number), which greatly clarifies
which marks denote chapters and which denote verses. Secondly, it fixes an issue
of marks not being properly calculated at the top of the first page.