sbtspaper: Move footnote config to end to avoid overruling.
My footnote specification to conform with SBTS style had at some point stopped
working, presumably because some other package or setting in the class file was
overriding it.

Has now been moved to near the end of the class file, and it works as expected
now.

Additionally, shorten the length of the footnote separating line to conform more
completely with SBTS style.