sbtspaper: ensure footnotes at bottom of page.
If the text on a page does not fill the whole page, LaTeX was placing the
footnote block directly below the main text block, with empty space below the
footnotes (i.e the footnotes were floating up the page). I think I had found a
way to avoid this on the very last page, but on every other page, the footnotes
floated up.

By adding \vfill to the definition of \footnoterule, which I was redefining
anyway to change the length to fit with SBTS requirements, the footnotes now sit
at the bottom of the page, with the appropriate gap between the main text block
and the footnotes block.

Also update date and version number of class definition to reflect recent
changes to footnote definitions (this and previous commits).