Update handout class to include use of Hebrew text (v1.2)
Required a significant restructuring of the class file as packages need to be
loaded *before* Hebrew font settings and bidi package are loaded, however the
loading of Hebrew fonts at the end of the class file overrides other settings,
such as \maketitle and contents page settings. So these are redefined after the
loading of Hebrew config.

It is possible that this means that my maketitle, section headings and contents
page settings will now break if Hebrew text is included, will need to
investigate the documentation if this is so.