Use microtype package when compiling adharticle documents with XeTeX
It is no longer the case that the excellent microtype package is incompatible
with XeLaTeX. Though not all features are available with XeLaTeX (possibly a
good reason to switch to using LuaLaTeX, but check speed issues), loading the
package without options will enable what microtypographical features are
compatible with the XeLaTeX system. So load the package and benefit from what is
accessible.